refactor: separate package for video math utils
